+++ src/Education.Client/Gateways/History/Topic/GetTopicsRequest.cs
@@ -3,7 +3,7 @@ using Education.Client.Gateways.Models;
 
 namespace Education.Client.Gateways.History.Topic;
 
-public sealed record GetTopicsRequest(EpochType Epoch, int Page, short Count)
+public sealed record GetTopicsRequest(EpochType Epoch, int Page, int Count)
     : PageRequest(Page, Count)
 {
     public override string ToQuery() =>
